refactor(main): extract components
