+++ packages/administration/navbar/index.ts
@@ -1,0 +1,2 @@
+export * from './xm-navbar.module';
+export * from './navbar.component';
