+++ packages/administration/navbar/navbar.component.html
