+++ packages/administration/navbar/navbar.component.scss
@@ -1,4 +1,4 @@
-@import 'theme';
+@import '@xm-ngx/styles/theming';
 
 
 // TODO; redesign and remove ng-deep
