+++ packages/administration/navbar/xm-navbar.module.ts
