+++ packages/components/navbar-heatmap-widget/heatmap-container.scss
@@ -1,10 +1,10 @@
-:host ::ng-deep .heatmap-container-wrapper {
+:host.heatmap-container-wrapper {
   width:    100%;
   height:   100%;
   position: absolute;
 }
 
-:host ::ng-deep #heatmapContainer {
+#heatmapContainer {
   width:  100%;
   height: 100%
 }
