+++ packages/components/navbar-heatmap-widget/index.ts
@@ -1,2 +1,3 @@
 export { NavbarHeatmapWidgetComponent, NavbarHeatmapWidgetModule } from './navbar-heatmap-widget.component';
 export { HeatmapService } from './heatmap.service';
+export { XmHeatmapContainerComponent } from './xm-heatmap-container.component';
