+++ packages/components/page-ribbon/index.ts
@@ -1,0 +1,1 @@
+export * from './page-ribbon.component';
