+++ packages/components/page-ribbon/page-ribbon.scss
@@ -1,6 +1,3 @@
-/* ==========================================================================
-Developement Ribbon
-========================================================================== */
 .ribbon {
     background-color: rgba(170, 0, 0, 0.5);
     left:             -3.5em;
