+++ packages/components/page-ribbon/profiles/index.ts
@@ -1,0 +1,2 @@
+export * from './profile-info.model';
+export * from './profile.service';
