+++ packages/components/page-ribbon/profiles/profile-info.model.ts
