--- src/app/layouts/footer/footer.component.ts
@@ -1,7 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-    selector: 'xm-footer',
-    templateUrl: './footer.component.html',
-})
-export class FooterComponent {}
