--- src/app/layouts/index.ts
@@ -1,6 +1,0 @@
-export * from './main/main.component';
-export * from './footer/footer.component';
-export * from './navbar/navbar.component';
-export * from './profiles/page-ribbon.component';
-export * from './profiles/profile.service';
-export * from './profiles/profile-info.model';
