+++ src/app/layouts/main/index.ts
@@ -1,0 +1,1 @@
+export * from './main.component';
